refactor hover actions for interactions settings
