+++ resources/js/components/Factory/Main/Interactions/ConsentInteraction.vue
